--- EssentialApp/EssentialAppTests/Helpers/FeedUIIntegrationTests+Localization.swift
@@ -1,25 +1,0 @@
-//
-//  Copyright © 2019 Essential Developer. All rights reserved.
-//
-
-import Foundation
-import XCTest
-import EssentialFeed
-
-extension FeedUIIntegrationTests {
-	private class DummyView: ResourceView {
-		func display(_ viewModel: Any) {}
-	}
-	
-	var loadError: String {
-		LoadResourcePresenter<Any,DummyView>.loadError
-	}
-	
-	var feedTitle: String {
-		FeedPresenter.title
-	}
-	
-	var commentsTitle: String {
-		ImageCommentsPresenter.title
-	}
-}
